fix: AnnouncementControllerTest fixed.
The problem was the endpoint. There are two methods that was pointed to the same endpoint one of them was  getByTitle and one of them was getById.

If you perform a request like ./announcement/1 ---> the compiler does not understand if 1 is the title or the id so it throws illegal argument exception.

Its fixed by requesting a query param. The changes must made accordingly in the frontend side.